requirements.txt
Django==1.9.1
MySQL-python==1.2.5
Pillow==3.1.0
Pygments==2.0.2
argparse==1.2.1
bcrypt==2.0.0
blessings==1.6
bpython==0.14.2
cffi==1.4.2
curtsies==0.1.23
django-appconf==1.0.1
django-braces==1.8.1
django-compressor==2.0
django-contrib-comments==1.6.2
django-debug-toolbar==1.4
django-geoposition==0.2.2
django-model-utils==2.4
django-wysiwyg-redactor==0.4.9
greenlet==0.4.9
gunicorn==19.4.5
logutils==0.3.3
pycparser==2.14
rcssmin==1.0.6
requests==2.9.1
rjsmin==1.0.12
sendgrid==1.5.20
six==1.10.0
smtpapi==0.2.0
sqlparse==0.1.18
wsgiref==0.1.2
